Bugfix: deleting cookies now sets the date far in the past
The delete method used to set the cookie again using the current date minus 1 hour. The
problem with this approach is that for some reasons local datetime settings are taken
into account. A better way is to make sure this date is far in the past to avoid any
mishaps. Therefor the number 1 (January 1st 1970) seemed logical to use.